refactor: resetInput 파일에 resetCarInput, resetCountInput 통합
