+++ docs/README.md
@@ -48,7 +48,7 @@
 - [x] alert는 void형이기 때문에, 삼항연산자 지양 
 - [x] hidden없애고 id로 하기 
 - [x] 상수 (하드코딩 x)
-- [x] alert파일 데이터 분리 
+- [ ] alert파일 데이터 분리 
 - [ ] 파일 정리해보기 (여러개인것 합치기)
 
 ## ✅ 요구 사항
