+++ src/js/handlers/handleCarNameInput.js
@@ -1,7 +1,7 @@
 import $ from '../utils/dom.js';
 import renderCountInput from '../views/renderCountInput.js';
 import { isOverFive, isDuplicate, alertCarInputError } from '../views/alertInputError.js';
-import resetCarInput from '../views/resetCarInput.js';
+import { resetCarInput } from '../views/resetInput.js';
 import { disableCarInput } from '../views/disableInput.js';
 import { CAR } from '../utils/constants.js';
