+++ src/js/handlers/handleRacingCountInput.js
@@ -1,7 +1,7 @@
 import $ from '../utils/dom.js';
 import { disableCountInput } from '../views/disableInput.js';
 import { alertCountInputError } from '../views/alertInputError.js';
-import resetCountInput from '../views/resetCountInput.js';
+import { resetCountInput } from '../views/resetInput.js';
 import playGame from './playGame.js';
 import { COUNT } from '../utils/constants.js';
