--- src/js/views/resetCarInput.js
@@ -1,7 +1,0 @@
-import $ from '../utils/dom.js';
-
-const resetCarInput = () => {
-  $('#car-names-input').value = '';
-};
-
-export default resetCarInput;
