--- src/js/views/resetCountInput.js
@@ -1,7 +1,0 @@
-import $ from '../utils/dom.js';
-
-const resetCountInput = () => {
-  $('#racing-count-input').value = '';
-};
-
-export default resetCountInput;
