+++ src/js/views/resetInput.js
@@ -1,0 +1,9 @@
+import $ from '../utils/dom.js';
+
+export const resetCarInput = () => {
+  $('#car-names-input').value = '';
+};
+
+export const resetCountInput = () => {
+  $('#racing-count-input').value = '';
+};
